REF: fixings all in same location (#123)
(cherry picked from commit c6f88d3990aa254b9f60e8d998c8186b748d0cb2)
