+++ python/rateslib/data/historical/eur_rfr.csv
