+++ python/rateslib/data/historical/sek_rfr.csv
