+++ python/rateslib/data/historical/swestr.csv
